[packaging] Change inline pkgconfig entry to fix build
tar_git looks for entries starting with "Name" in the .spec, sorts
the result and uses the first line as name of the tarball that it
will create and send to OBS.

Because of the sorting step, tar_git was using "id3tag" instead of
libid3tag as name of the tarball, resulting in a build failure
because the build script could not find the tarball holding the
source.

This commit changes the name of the inline pkgconfig file
(which is just an indication anyway) to prevent that confusion
from happening.

It is, of course, just a workaround.